+++ app/cards/[id]/edit/page.tsx
@@ -1,4 +1,4 @@
-import EditDeckCards from "@/src/deck/EditDeckCards"
+import EditCards from "@/src/card/edit/EditCards"
 import { getDeckDetail } from "@/src/deck/api"
 import { getQueryClient } from "@/src/shared/get-query-client"
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query"
@@ -12,7 +12,7 @@ export default async function page({ params }: { params: { id: number } }) {
   }
   return (
     <HydrationBoundary state={dehydrate(queryClient)}>
-      <EditDeckCards />
+      <EditCards />
     </HydrationBoundary>
   )
 }
