--- src/deck/EditDeckCards.tsx
@@ -1,445 +1,0 @@
-"use client"
-
-import { Button } from "@/src/shadcn/components/ui/button"
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-} from "@/src/shadcn/components/ui/form"
-import { Input } from "@/src/shadcn/components/ui/input"
-import { useFieldArray, useForm } from "react-hook-form"
-import { getDeckDetail } from "@/src/deck/api"
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
-import { useParams } from "next/navigation"
-import { Card, CardHeader, CardTitle } from "@/src/shadcn/components/ui/card"
-import { ImageUp, Mic, X } from "lucide-react"
-import { Textarea } from "@/src/shadcn/components/ui/textarea"
-import { mutateCards } from "@/src/card/api"
-import { FormEvent, useRef, useState } from "react"
-import { ScrollArea } from "@/src/shadcn/components/ui/scroll-area"
-import { toast } from "@/src/shadcn/hooks/use-toast"
-import Image from "next/image"
-
-const createDummyFile = (attachment?: Attachment) => {
-  const file = new File([], attachment?.fileName ?? "", {
-    type: attachment?.fileType,
-  })
-  return file
-}
-
-const preventEnterKeySubmission = (e: React.KeyboardEvent<HTMLFormElement>) => {
-  const target = e.target
-  if (e.key === "Enter" && target instanceof HTMLInputElement) {
-    e.preventDefault()
-  }
-}
-const toCardForRequestValue = (value: CardForForm[keyof CardForForm]) => {
-  if (typeof value === "number") return value.toString()
-  if (typeof value === "string") return value
-  if (value instanceof File) return value
-  return value
-}
-const getFileNameFromField = (file: File | string | undefined) => {
-  if (file instanceof File) return file.name
-  return file
-}
-const getUpdatedFields = (
-  card: CardForForm,
-  dirtyFields: {
-    [key in keyof CardForEditRequest]?: boolean
-  },
-) => {
-  const keys = Object.keys(card) as Array<keyof CardForEditRequest>
-  return keys.reduce((acc, key) => {
-    const isRequiredKey = key === "cardId" || key === "deckId"
-    if (dirtyFields[key] || isRequiredKey)
-      return { ...acc, [key]: toCardForRequestValue(card[key]) }
-    return acc
-  }, {})
-}
-export default function DeckEdit() {
-  const { id } = useParams<{ id: string }>() as unknown as { id: number }
-  const queryClient = useQueryClient()
-  const { data } = useQuery(getDeckDetail(id))
-  const fileInputs = useRef<Map<string, HTMLElement> | null>(null)
-  const [lastTouchedCard, setTouchedCard] = useState<CardForForm>()
-  const cards: CardForForm[] =
-    data?.cards.map(({ attachments, id, ...card }) => ({
-      ...card,
-      cardId: id.toString(),
-      image:
-        attachments.find((v) => v.fileType.startsWith("image"))?.fileName ?? "",
-      audio:
-        attachments.find((v) => v.fileType.startsWith("audio"))?.fileName ?? "",
-      status: "DEFAULT",
-      attachments: attachments,
-    })) ?? []
-  const form = useForm({ values: { cards } })
-  const { fields, update, append } = useFieldArray({
-    control: form.control,
-    name: "cards",
-    keyName: "key",
-  })
-  const cardMutation = useMutation({
-    mutationFn: mutateCards,
-    onSuccess: () => {
-      queryClient.invalidateQueries(getDeckDetail(id))
-    },
-    onError: () => {
-      toast({
-        title: `Failed update cards`,
-        variant: "destructive",
-      })
-    },
-  })
-  const onSubmit = (data: { cards: CardForForm[] }) => {
-    const dirtyCards = form.formState.dirtyFields.cards
-    if (!dirtyCards?.length) return
-    const result = data.cards.reduce(
-      (acc, card, i) => {
-        if (!dirtyCards[i]) return acc
-        const { key, ...dirtyFields } = dirtyCards[i]
-        const isDeleted = card.status === "DELETE" && card.cardId !== undefined
-        const isEdited =
-          Object.values(dirtyFields).some((v) => v) && card.status !== "DELETE"
-        if (isDeleted) acc.deleted.cardIds.push(card.cardId!)
-        if (isEdited) acc.edited.push(getUpdatedFields(card, dirtyFields))
-        return acc
-      },
-      {
-        deleted: { deckId: id, cardIds: [] as string[] },
-        edited: [] as CardForEditRequest[],
-      },
-    )
-    if (result.edited.length || result.deleted.cardIds.length)
-      cardMutation.mutate(result)
-  }
-  if (!data) return <div>loading...</div>
-  const handleAddCard = () => {
-    append({
-      name: "",
-      description: "",
-      deckId: id,
-      status: "NEW",
-      image: createDummyFile(),
-      audio: createDummyFile(),
-      attachments: [],
-    })
-  }
-  const handleRemove = (index: number) => () => {
-    update(index, {
-      ...fields[index],
-      status: "DELETE",
-    })
-    form.trigger()
-  }
-  const handleUndo = (index: number) => () => {
-    if (!fields[index].cardId) {
-      update(index, { ...fields[index], status: "NEW" })
-      form.trigger()
-      return
-    }
-    update(index, { ...fields[index], status: "DEFAULT" })
-  }
-  const handleLastTouchedCard = (event: FormEvent) => {
-    if (event.target instanceof HTMLElement) {
-      const element = event.target.closest("div[data-index]")
-      if (element instanceof HTMLElement) {
-        setTouchedCard(
-          form.getValues().cards[element.dataset.index as unknown as number],
-        )
-        form.trigger()
-      }
-    }
-  }
-  const setFileInputNode = (key: string) => (node: HTMLInputElement | null) => {
-    const map = getFileInputNode()
-    map.set(key, node!)
-  }
-
-  const getFileInputNode = () => {
-    if (!fileInputs.current) {
-      fileInputs.current = new Map()
-    }
-    return fileInputs.current
-  }
-  return (
-    <Form {...form}>
-      <form
-        onChange={handleLastTouchedCard}
-        onFocus={handleLastTouchedCard}
-        onClick={handleLastTouchedCard}
-        onSubmit={form.handleSubmit(onSubmit)}
-        onKeyDown={preventEnterKeySubmission}
-        className="w-full flex flex-col flex-grow overflow-auto"
-      >
-        <div className="grid grid-cols-1 md:grid-cols-2 flex-1 overflow-auto">
-          {/* scrollArea 내부적으로 사용하는 display table로 인해 text-overflow가 정상 동작 하지 않아 수정. */}
-          <ScrollArea
-            className="w-full min-w-0 [&>[data-radix-scroll-area-viewport]>div]:!block border-r border-gray-200"
-            type="auto"
-          >
-            <ul className="flex flex-col w-full overflow-auto justify-items-center p-3">
-              {fields.map((card, index) => {
-                const isDeleted = card.status === "DELETE"
-                return (
-                  <li key={card.key} className="flex justify-center ">
-                    <Card
-                      className="flex flex-col p-2 m-2 gap-2 max-w-md w-full flex-shrink-0"
-                      data-index={index}
-                    >
-                      <div className="flex flex-col gap-2">
-                        <FormField
-                          name={`cards.${index}.name`}
-                          control={form.control}
-                          render={({ field }) => (
-                            <FormItem>
-                              <FormLabel>name</FormLabel>
-                              <FormControl>
-                                <Input
-                                  disabled={isDeleted}
-                                  type="text"
-                                  placeholder="name"
-                                  {...form.register(field.name, {
-                                    required: !isDeleted,
-                                  })}
-                                  value={field.value ?? card.name}
-                                />
-                              </FormControl>
-                            </FormItem>
-                          )}
-                        ></FormField>
-                        <FormField
-                          name={`cards.${index}.description`}
-                          control={form.control}
-                          render={({ field }) => (
-                            <FormItem>
-                              <FormLabel>description</FormLabel>
-                              <FormControl>
-                                <Textarea
-                                  disabled={isDeleted}
-                                  rows={5}
-                                  placeholder="description"
-                                  {...form.register(field.name, {
-                                    required: !isDeleted,
-                                  })}
-                                  value={field.value ?? card.description}
-                                />
-                              </FormControl>
-                            </FormItem>
-                          )}
-                        ></FormField>
-                      </div>
-                      <div className="flex justify-between ">
-                        <div className="flex flex-col gap-1 min-w-0">
-                          <FormField
-                            name={`cards.${index}.image`}
-                            render={({ field: { value, onChange, name } }) => (
-                              <FormItem>
-                                <div className="text-ellipsis flex items-center gap-1">
-                                  <Button
-                                    aria-label="upload-image"
-                                    disabled={isDeleted}
-                                    type="button"
-                                    variant={"outline"}
-                                    size={"icon"}
-                                    onClick={() => {
-                                      getFileInputNode().get(name)?.click()
-                                    }}
-                                  >
-                                    <ImageUp />
-                                  </Button>
-                                  <FormLabel className="truncate cursor-pointer flex-1 leading-normal">
-                                    {getFileNameFromField(value ?? card.image)}
-                                  </FormLabel>
-                                  <Button
-                                    aria-label="remove-image"
-                                    className={
-                                      getFileNameFromField(value ?? card.image)
-                                        ?.length
-                                        ? undefined
-                                        : "hidden"
-                                    }
-                                    disabled={isDeleted}
-                                    type="reset"
-                                    variant={"ghost"}
-                                    size={"icon"}
-                                    onClick={() => {
-                                      onChange(createDummyFile())
-                                    }}
-                                  >
-                                    <X />
-                                  </Button>
-                                </div>
-                                <FormControl>
-                                  <Input
-                                    disabled={isDeleted}
-                                    type="file"
-                                    className="hidden"
-                                    accept="image/*"
-                                    {...form.register(name)}
-                                    ref={setFileInputNode(name)}
-                                    onChange={(e) => {
-                                      onChange(e.target.files?.[0])
-                                    }}
-                                  />
-                                </FormControl>
-                              </FormItem>
-                            )}
-                          ></FormField>
-                          <FormField
-                            name={`cards.${index}.audio`}
-                            render={({ field: { onChange, value, name } }) => (
-                              <FormItem>
-                                <div className="text-ellipsis flex items-center gap-1">
-                                  <Button
-                                    aria-label="upload-audio"
-                                    disabled={isDeleted}
-                                    type="button"
-                                    variant={"outline"}
-                                    size={"icon"}
-                                    onClick={() => {
-                                      getFileInputNode().get(name)?.click()
-                                    }}
-                                  >
-                                    <Mic />
-                                  </Button>
-                                  <FormLabel className="truncate flex-1 cursor-pointer leading-normal">
-                                    {getFileNameFromField(value ?? card.audio)}
-                                  </FormLabel>
-                                  <Button
-                                    aria-label="remove-audio"
-                                    className={
-                                      getFileNameFromField(value ?? card.audio)
-                                        ?.length
-                                        ? undefined
-                                        : "hidden"
-                                    }
-                                    disabled={isDeleted}
-                                    type="reset"
-                                    variant={"ghost"}
-                                    size={"icon"}
-                                    onClick={() => {
-                                      onChange(createDummyFile())
-                                    }}
-                                  >
-                                    <X />
-                                  </Button>
-                                  <FormControl>
-                                    <Input
-                                      disabled={isDeleted}
-                                      type="file"
-                                      className="hidden"
-                                      accept="audio/*"
-                                      {...form.register(name)}
-                                      ref={setFileInputNode(name)}
-                                      onChange={(e) =>
-                                        e.target.files &&
-                                        onChange(e.target.files[0])
-                                      }
-                                    />
-                                  </FormControl>
-                                </div>
-                              </FormItem>
-                            )}
-                          ></FormField>
-                        </div>
-                        {isDeleted ? (
-                          <Button
-                            type="button"
-                            onClick={handleUndo(index)}
-                            variant={"secondary"}
-                            size={"sm"}
-                          >
-                            undo
-                          </Button>
-                        ) : (
-                          <Button
-                            type="button"
-                            onClick={handleRemove(index)}
-                            variant={"destructive"}
-                            size={"sm"}
-                          >
-                            remove
-                          </Button>
-                        )}
-                      </div>
-                    </Card>
-                  </li>
-                )
-              })}
-            </ul>
-          </ScrollArea>
-          <div className="hidden md:flex w-full truncate items-center justify-center px-2">
-            <CardPreview props={lastTouchedCard ?? cards[0]} />
-          </div>
-        </div>
-        <div className="flex justify-around p-3 border-t">
-          <Button type="button" onClick={handleAddCard} variant={"outline"}>
-            Add card
-          </Button>
-          <Button type="submit">Submit</Button>
-        </div>
-      </form>
-    </Form>
-  )
-}
-
-const getFileSrc = (
-  type: "audio" | "image",
-  attachments: Attachment[],
-  file?: File | string,
-) => {
-  if (file instanceof File && file.size) return URL.createObjectURL(file)
-  // already uploaded file
-  if (typeof file === "string") {
-    return attachments.find((v) => v.fileType.startsWith(type))?.filePath ?? ""
-  }
-  return ""
-}
-function CardPreview({ props }: { props?: CardForForm }) {
-  if (!props) return <div></div>
-  const imageSrc = getFileSrc("image", props.attachments, props.image)
-  const audioSrc = getFileSrc("audio", props.attachments, props.audio)
-  const hasImage = !!props.image && !!imageSrc.length
-  const hasAudio = !!props.audio && !!audioSrc.length
-  return (
-    <Card
-      className="
-      w-full
-      break-words whitespace-normal 
-      overflow-hidden
-      max-w-md flex flex-col h-[30rem]"
-    >
-      <div className="basis-1/2 object-contain w-full">
-        {hasImage ? (
-          <Image
-            src={imageSrc}
-            alt="card image"
-            width={0}
-            height={0}
-            className="h-[15rem] object-contain w-full"
-          ></Image>
-        ) : (
-          <div className="w-full h-full text-center content-center bg-slate-100 text-gray-700">
-            No Image
-          </div>
-        )}
-      </div>
-
-      <CardHeader className="basis-1/2 m-0 overflow-auto border-t h-full">
-        <CardTitle className="text-lg">{props.name}</CardTitle>
-        <ScrollArea type="auto" className="h-full">
-          <div className="overflow-auto whitespace-break-spaces break-all">
-            {props.description}
-          </div>
-        </ScrollArea>
-        {hasAudio && (
-          <audio className="self-center" controls src={audioSrc}></audio>
-        )}
-      </CardHeader>
-    </Card>
-  )
-}
